fix: namespaces changes
