+++ config/default/QtiItemContentTokenizer.conf.php
@@ -3,7 +3,7 @@
  * Default config header created during install
  */
 
-use oat\taoQtiItem\model\search\Tokenizer\Filter\HasValueFilter;
+use oat\taoQtiItem\model\search\Tokenizer\Filter\ClearValueFilter;
 use oat\taoQtiItem\model\search\Tokenizer\Filter\NotBase64ContentFilter;
 use oat\taoQtiItem\model\search\Tokenizer\Filter\NotJsonFilter;
 use oat\taoQtiItem\model\search\QtiItemContentTokenizer;
@@ -12,7 +12,7 @@ return new oat\taoQtiItem\model\search\QtiItemContentTokenizer(
     [
         QtiItemContentTokenizer::OPTION_FILTERS => [
             new NotJsonFilter(),
-            new HasValueFilter(),
+            new ClearValueFilter(),
             new NotBase64ContentFilter(),
         ]
     ]
